--- tests/imagery/test_data/test_feature.json
@@ -1,31 +1,0 @@
-{
-    "type": "Feature",
-    "properties": {},
-    "geometry": {
-        "coordinates": [
-            [
-                [
-                    -122.44042073493674,
-                    37.78393485513989
-                ],
-                [
-                    -122.44042073493674,
-                    37.78258944782999
-                ],
-                [
-                    -122.43841053975626,
-                    37.78258944782999
-                ],
-                [
-                    -122.43841053975626,
-                    37.78393485513989
-                ],
-                [
-                    -122.44042073493674,
-                    37.78393485513989
-                ]
-            ]
-        ],
-        "type": "Polygon"
-    }
-}
